% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-centroids.R
\docType{data}
\name{centroids}
\alias{centroids}
\alias{CED_2016_centroids}
\alias{LGA_2016_centroids}
\alias{POA_2016_centroids}
\alias{SA1_2011_centroids}
\alias{SA1_2016_centroids}
\alias{SA2_2011_centroids}
\alias{SA2_2011_weighted_centroids}
\title{Centroids}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 150 rows and 3 columns.}
\usage{
CED_2016_centroids

LGA_2016_centroids

POA_2016_centroids

SA1_2011_centroids

SA1_2016_centroids

SA2_2011_centroids

SA2_2011_weighted_centroids
}
\description{
Centroids of the corresponding shapefile.
}
\keyword{datasets}
