% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grattan_leaflet.R
\name{grattan_leaflet}
\alias{grattan_leaflet}
\title{Draw leaflet charts in grattan style}
\usage{
grattan_leaflet(DT, Year = c("2011", "2016"))
}
\arguments{
\item{DT}{A data frame whose first column is of the form \code{SA[1-4]_(CODE|NAME)(11|16)} as a key to join with ASGS shapefiles.
Other required names are \code{fillColor}, \code{labelTitle}, and \code{labelText}.}

\item{Year}{The year to which \code{DT} applies.}
}
\value{
A \code{leaflet} object; a map of Australia.
}
\description{
Draw leaflet charts in grattan style
}
