% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlon2SA.R
\name{latlon2SA}
\alias{latlon2SA}
\title{Decode coordinates to Statistical Areas}
\usage{
latlon2SA(lat, lon, to = c("SA2", "SA1", "SA3", "SA4"), return = c("sp",
  "v"))
}
\arguments{
\item{lat, lon}{The latitude and longitude of the points.}

\item{to}{Which statistical areas (by default SA2) to map to.}

\item{return}{Return the spatial data frame or a vector?}
}
\value{
A vector of the SAs mapped.
}
\description{
Decode coordinates to Statistical Areas
}
