% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Postcode.R
\docType{data}
\name{Postcode}
\alias{Postcode}
\title{Postcode shapefile}
\format{An object of class \code{SpatialPolygonsDataFrame} with 2668 rows and 3 columns.}
\usage{
Postcode
}
\description{
The shapefile of 2016 postcodes.
}
\keyword{datasets}
