% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RF.R
\name{RF}
\alias{RF}
\title{RF}
\usage{
RF(X, Y, r)
}
\arguments{
\item{X}{X is a matrix of marker genotype of size n×p where n are no of Individuals under study (i.e. genotype, lines) and p are number of markers.}

\item{Y}{Y is a vector of individuals of size n×1.}

\item{r}{fraction of testing data (ranges from (0 to 1)) used during model fitting (suppose if one want to use 0.9 of data for model training and remaining 0.1 for model testing so one has to define r=0.1).}
}
\value{
$fit RF model fitting
       $Accuracy prediction accuracy of the testing population (pearson correlation coefficient between actual phenotypic value and predicted phenotypic value)
}
\description{
Calculates the Genomic Estimated Breeding Value based on RF method.
}
\details{
This function fits model by dividing data into two part i.e. training population and testing population. Former one is used to build the models and later one for performance evaluation. The performance of model is evaluated by calculating model's prediction accuracy of testing population (pearson correlation coefficient between actual phenotypic value and predicted phenotypic value). Whole procedures is repeated 50 times and accuracy is averaged. For machine learning methods, an auto hyperparameter adjusting through grid search is implemented to achieve better prediction accuracy. In each split, the 0.9 of the training population is used to adjust hyperparameters, of which 0.9 selected to be training sets and 0.1 to be testing sets, 10 times of simple validation is implemented.
}
\examples{
library(ASGS);rm (list=ls ());data(trout);
X=trout[1:50,2:11];Y=as.data.frame(trout[1:50,1]);r=0.1;ASGS::RF(X,Y,r)
}
