% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA2016_decoder.R
\docType{data}
\name{SA2016_decoder}
\alias{SA2016_decoder}
\title{Decode statistical areas}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 2292 rows and 13 columns.
}
\usage{
SA2016_decoder
}
\description{
A \code{data.table} showing which higher-level statistical area each \code{SA2} is within.
}
\keyword{datasets}
