% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAs-simple.R
\docType{data}
\name{SA3_2011_simple}
\alias{SA3_2011_simple}
\title{Simplified shapefiles}
\format{An object of class \code{SpatialPolygonsDataFrame} with 333 rows and 10 columns.}
\usage{
SA3_2011_simple
}
\description{
Simpler (less precision, fewer vertices) of the \code{SA[1-4]} shapefiles.
Smaller so more suitable for some plots.
}
\keyword{datasets}
