% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAs-simple.R
\docType{data}
\name{SA_simple}
\alias{SA_simple}
\alias{SA3_2011_simple}
\alias{SA3_2016_simple}
\alias{SA1_2011_simple}
\alias{SA2_2011_simple}
\alias{SA2_2016_simple}
\title{Simplified shapefiles}
\format{
An object of class \code{function} of length 1.
}
\usage{
SA3_2011_simple()

SA3_2016_simple()

SA1_2011_simple()

SA2_2011_simple()

SA2_2016_simple()

SA3_2016_simple
}
\description{
Simpler (less precision, fewer vertices) of the \code{SA[1-4]} shapefiles.
Smaller so more suitable for some plots.
}
\keyword{datasets}
