% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_shapefile.R
\name{smooth_shapefile}
\alias{smooth_shapefile}
\title{Smooth a shapfile by nearest neighbourhood}
\usage{
smooth_shapefile(shapefile, data_by_id, id, var2smooth, k = 10,
  coalesce.to = 0)
}
\arguments{
\item{shapefile}{A shapefile (as from this package) containing an \code{as.data.frame} method.}

\item{data_by_id}{A \code{data.frame} containing a value for each polygon within \code{shapefile}.}

\item{id}{The key of \code{data_by_id} on which the shapefile may be merged with \code{shapefile}.}

\item{var2smooth}{A length-one character vector indicating which variable in \code{data_by_id} to smooth.}

\item{k}{Passed to \code{\link[spdep]{knearneigh}}.}

\item{coalesce.to}{A single value. If there are any missing values in \code{data_by_id[[id]]}, what should these be imputed as? Must
have the same storage mode as \code{data_by_id[[id]]}.}
}
\value{
A vector of values representing the smoothed version of \code{data_by_id[[id]]}.
}
\description{
Given a shapefile and a \code{data.frame} with a value for each polygon,
return a smoothed value of the values.
}
