% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAs.R
\docType{data}
\name{SA1}
\alias{SA1}
\alias{SA2_2011}
\alias{SA1_2011}
\title{SA2}
\format{An object of class \code{SpatialPolygonsDataFrame} with 2196 rows and 12 columns.}
\usage{
SA2_2011

SA1_2011
}
\description{
ESRI shape files.
}
\keyword{datasets}
