% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-centroids.R
\docType{data}
\name{centroids}
\alias{centroids}
\alias{CED_2016_centroids}
\alias{LGA_2016_centroids}
\alias{POA_2016_centroids}
\alias{SA1_2011_centroids}
\alias{SA1_2016_centroids}
\alias{SA2_2011_centroids}
\alias{SA2_2011_weighted_centroids}
\alias{cd06act_centroid}
\alias{cd06nsw_centroid}
\alias{cd06nt_centroid}
\alias{cd06ot_centroid}
\alias{cd06qld_centroid}
\alias{cd06sa_centroid}
\alias{cd06tas_centroid}
\alias{cd06vic_centroid}
\alias{cd06wa_centroid}
\alias{sla06act_centroid}
\alias{sla06nsw_centroid}
\alias{sla06nt_centroid}
\alias{sla06ot_centroid}
\alias{sla06qld_centroid}
\alias{sla06sa_centroid}
\alias{sla06tas_centroid}
\alias{sla06vic_centroid}
\alias{sla06wa_centroid}
\title{Centroids}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 150 rows and 3 columns.}
\usage{
CED_2016_centroids

LGA_2016_centroids

POA_2016_centroids

SA1_2011_centroids

SA1_2016_centroids

SA2_2011_centroids

SA2_2011_weighted_centroids

cd06act_centroid

cd06nsw_centroid

cd06nt_centroid

cd06ot_centroid

cd06qld_centroid

cd06sa_centroid

cd06tas_centroid

cd06vic_centroid

cd06wa_centroid

sla06act_centroid

sla06nsw_centroid

sla06nt_centroid

sla06ot_centroid

sla06qld_centroid

sla06sa_centroid

sla06tas_centroid

sla06vic_centroid

sla06wa_centroid
}
\description{
Centroids of the corresponding shapefile.
}
\keyword{datasets}
