% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grattan_leaflet.R
\name{grattan_leaflet}
\alias{grattan_leaflet}
\title{Draw leaflet charts in grattan style}
\usage{
grattan_leaflet(DT, Year = c("2011", "2013", "2016"), simple = FALSE,
  na.value = "#6A737B")
}
\arguments{
\item{DT}{A data frame whose first column is of the form \code{SA[1-4]_(CODE|NAME)(11|16)} as a key to join with ASGS shapefiles, or \code{CED} for Commonwealth Electoral Divisions.
Other required names are \code{fillColor}, \code{labelTitle}, and \code{labelText}.}

\item{Year}{The year to which \code{DT} applies. May be 2011 or 2016 for \code{SA[1-4]} shapefiles or 2013 or 2016 for \code{CED}.}

\item{simple}{Use a simplified shapefile (if available)?}

\item{na.value}{\code{fillColor} to use for unmatched polygons.}
}
\value{
A \code{leaflet} object; a map of Australia.
}
\description{
Draw leaflet charts in grattan style
}
\examples{
\dontrun{
library(readxl)
library(data.table)
library(magrittr)
library(grattanCharts)

DT <- 
  read_excel(system.file("extdata", "unemp.xlsx", package = "ASGS")) \%>\%
  setDT \%>\%
  .[, .(SA2_NAME11 = `Statistical Area Level 2 (SA2)`,
        fillColor = gpal(7)[Breaks], 
        labelText = `Statistical Area Level 2 (SA2)`,
        labelTitle = "Unemployment")]
        
grattan_leaflet(DT, Year = "2011")
}


}
