% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non-SAs.R
\name{non-ABS-shapefile}
\alias{non-ABS-shapefile}
\alias{CED_2016}
\alias{CED_2013}
\alias{LGA_2016}
\alias{POA_2016}
\alias{POA_2021}
\alias{LGA_2021}
\alias{LGA_2022}
\alias{CED_2021}
\alias{DZN_2021}
\alias{SED_2021}
\alias{SED_2022}
\alias{TR_2021}
\alias{ADD_2021}
\alias{SAL_2021}
\title{Non-ASGS shapefiles}
\usage{
CED_2016()

CED_2013()

LGA_2016()

POA_2016()

POA_2021()

LGA_2021()

LGA_2022()

CED_2021()

DZN_2021()

SED_2021()

SED_2022()

TR_2021()

ADD_2021()

SAL_2021()
}
\description{
These are not the core Statistical Areas used by the ABS, 
but the areas are provided for convenience. CED (Commonwealth Electoral Divisions), 
LGA (Local Government Areas), POA (Postcodes).
}
