% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAs.R
\docType{data}
\name{SA1}
\alias{SA1}
\alias{SA1_2011}
\alias{SA2_2011}
\alias{SA3_2011}
\alias{SA1_2016}
\alias{SA2_2016}
\alias{SA3_2016}
\alias{SA4_2016}
\alias{SA4_2011}
\alias{DZN_2011}
\alias{DZN_2016}
\title{SA2}
\format{An object of class \code{SpatialPolygonsDataFrame} with 54772 rows and 14 columns.}
\usage{
SA1_2011

SA2_2011

SA3_2011

SA1_2016

SA2_2016

SA3_2016

SA4_2016

SA4_2011

DZN_2011

DZN_2016
}
\description{
ESRI shape files for statistical areas and destination zones.
}
\keyword{datasets}
