% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAs.R
\name{SA1}
\alias{SA1}
\alias{SA1_2011}
\alias{SA2_2011}
\alias{SA3_2011}
\alias{SA1_2016}
\alias{SA2_2016}
\alias{SA3_2016}
\alias{SA4_2016}
\alias{SA4_2011}
\alias{DZN_2011}
\alias{DZN_2016}
\alias{SA1_2021}
\alias{SA2_2021}
\alias{SA3_2021}
\alias{SA4_2021}
\alias{MB_2021}
\title{SA2}
\usage{
SA1_2011()

SA2_2011()

SA3_2011()

SA1_2016()

SA2_2016()

SA3_2016()

SA4_2016()

SA4_2011()

DZN_2011()

DZN_2016()

SA1_2021()

SA2_2021()

SA3_2021()

SA4_2021()

MB_2021()
}
\description{
\code{sf} objects for statistical areas
}
